+++ packages/io-auth-n-identity-domain/src/domain/value-objects/index.ts
@@ -1,2 +1,3 @@
 export * from "./common/index.js";
 export * from "./lollipop/index.js";
+export * from "./profile/index.js";
