+++ packages/io-auth-n-identity-domain/src/domain/value-objects/profile/index.ts
@@ -1,0 +1,1 @@
+export * from "./profile.value-object.js";
